backend/requirements.txt
fastapi
uvicorn[standard]
pandas
python-multipart
sentence-transformers
faiss-cpu
numpy
openai
python-dotenv
